Mudanças Church Lat e Long para double

+++ app/src/main/java/com/example/helio/gotochurchmobileproject/Basic/Church.java
@@ -15,8 +15,8 @@ public class Church implements Serializable{
     private Address address;
     private String adjunct;
     private int resIdImagem;
-    private long lat;
-    private long lng;
+    private double lat;
+    private double lng;
 
     public String getName() {
         return name;
@@ -66,19 +66,19 @@ public class Church implements Serializable{
         this.id = id;
     }
 
-    public long getLat() {
+    public double getLat() {
         return lat;
     }
 
-    public void setLat(long lat) {
+    public void setLat(double lat) {
         this.lat = lat;
     }
 
-    public long getLng() {
+    public double getLng() {
         return lng;
     }
 
-    public void setLng(long lng) {
+    public void setLng(double lng) {
         this.lng = lng;
     }
 }
